@@ -1,6 +1,6 @@
 name: flutter_incall_manager
 description: Flutter InCall Manager plugin for iOS/Android.
-version: 0.1.0
+version: 1.0.0+1
 homepage: https://github.com/kangshaojun/flutter-incall-manager
 
 dependencies:
@@ -16,5 +16,8 @@ flutter:
       ios:
         pluginClass: FlutterIncallManagerPlugin
 
+# environment:
+#   sdk: ">=2.0.0-dev.68.0 <3.0.0"
+
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
