@@ -22,7 +22,7 @@ class IncallManager {
   Future<void> start(
       {bool auto = true,
       MediaType media = MediaType.AUDIO,
-      String ringback}) async {
+      String? ringback}) async {
     await _channel.invokeMethod('start', <String, dynamic>{
       'media': media == MediaType.AUDIO ? 'audio' : 'video',
       'auto': auto,
@@ -31,7 +31,7 @@ class IncallManager {
   }
 
   /// Stop InCallManager
-  Future<void> stop({String busytone}) async {
+  Future<void> stop({String? busytone}) async {
     await _channel
         .invokeMethod('stop', <String, dynamic>{'busytone': busytone});
   }
